# docs/operations/iam-api.openapi.yaml
openapi: 3.1.0
info:
  title: IMS IAM API
  version: 1.0.0
  description: OpenAPI 3.1 contract for the IMS Identity & Access Management `/api/v1` surface.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [users]
      summary: Search users
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/roleId'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersEnvelope'
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
    post:
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
  /users/{id}:
    get:
      tags: [users]
      summary: Get user by id
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
    put:
      tags: [users]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
...
  /reports/iam/export-jobs/{jobId}/download:
    get:
      tags: [reports]
      summary: Download export file
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: File download
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    jobId:
      name: jobId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    search:
      name: search
      in: query
      schema: { type: string }
    status:
      name: status
      in: query
      schema: { type: string }
    branchId:
      name: branchId
      in: query
      schema: { type: string, format: uuid }
    roleId:
      name: roleId
      in: query
      schema: { type: string, format: uuid }
  responses:
    Problem400:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem401:
      description: Authentication error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem403:
      description: Authorization error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, errorCode]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errorCode: { type: string }
        correlationId: { type: string, nullable: true }
...
    CreateUserRequest:
      type: object
      additionalProperties: true
    UpdateUserRequest:
      type: object
      additionalProperties: true
...
    UserEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            user: { type: object, additionalProperties: true }
    PaginatedUsersEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items: { type: object, additionalProperties: true }
            total: { type: integer }
...